Fix typos, remove redundancies, and update logic for clarity
Corrected typos such as "Manu" to "Menu" and "Al" to "All" across classes. Replaced unnecessary variable declarations and improved method logic for better readability and functionality. Enhanced input validation, cleaned up dead code, and added more precise conditions where needed.

+++ service/src/main/java/greencity/ubstelegrambot/service/TelegramAuthorizationServiceImpl.java
@@ -55,8 +55,10 @@ public class TelegramAuthorizationServiceImpl implements TelegramAuthorizationSe
             Optional<UnknownTelegramUser> unknownSavedTelegramUser =
                 unknownTelegramUserRepository.findById(tgUserId);
             if (unknownSavedTelegramUser.isPresent()) {
-                unknownTelegramUserRepository.delete(unknownSavedTelegramUser.get());
-                user.ifPresent(value -> telegramBotRepository.save(createTelegramBotEntity(value, tgUserId, false)));
+                if (user.isPresent()) {
+                    telegramBotRepository.save(createTelegramBotEntity(user.get(), tgUserId, false));
+                    unknownTelegramUserRepository.delete(unknownSavedTelegramUser.get());
+                }
                 return TelegramUser.UNKNOWN_USER;
             }
